requirements.txt
demjson==2.2.4
requests==2.28.0
retry==0.9.2
skylark==0.9.1
PyGithub